# Optimize procfs/sysfs relative path creation