--- a/pkg/security/probe/probe.go
+++ b/pkg/security/probe/probe.go
@@ -190,7 +190,7 @@ func (p *Probe) VerifyEnvironment() *multierror.Error {
 			err = multierror.Append(err, errors.New("/etc/group doesn't seem to be a mountpoint"))
 		}
 
-		if mounted, _ := mountinfo.Mounted(util.HostProc()); !mounted {
+		if mounted, _ := mountinfo.Mounted(util.HostProc); !mounted {
 			err = multierror.Append(err, errors.New("/etc/group doesn't seem to be a mountpoint"))
 		}
 
@@ -198,7 +198,7 @@ func (p *Probe) VerifyEnvironment() *multierror.Error {
 			err = multierror.Append(err, fmt.Errorf("%s doesn't seem to be a mountpoint", p.kernelVersion.OsReleasePath))
 		}
 
-		securityFSPath := filepath.Join(util.GetSysRoot(), "kernel/security")
+		securityFSPath := filepath.Join(util.HostSys, "kernel/security")
 		if mounted, _ := mountinfo.Mounted(securityFSPath); !mounted {
 			err = multierror.Append(err, fmt.Errorf("%s doesn't seem to be a mountpoint", securityFSPath))
 		}