--- a/pkg/util/kernel/ip.go
+++ b/pkg/util/kernel/ip.go
@@ -17,6 +17,6 @@ import (
 
 // IsIPv6Enabled returns whether or not IPv6 has been enabled on the host
 func IsIPv6Enabled() bool {
-	_, err := ioutil.ReadFile(filepath.Join(util.GetProcRoot(), "net/if_inet6"))
+	_, err := ioutil.ReadFile(filepath.Join(util.HostProc, "net/if_inet6"))
 	return err == nil
 }